fix(oidc): client id not in audience whitelist
This fixes an issue where the audience whitelist did not contain the client id automatically.
