+++ docs/content/en/configuration/identity-providers/openid-connect/clients.md
@@ -28,14 +28,15 @@ intended for production use it's used to provide context and an indentation exam
 identity_providers:
   oidc:
     clients:
-      - id: 'myapp'
+      - id: 'unique-client-identifier'
         description: 'My Application'
         secret: '$pbkdf2-sha512$310000$c8p78n7pUMln0jzvd4aK4Q$JNRBzwAo0ek5qKn50cFzzvE9RXV88h1wJn5KGiHrD0YKtZaR/nCb2CJPOsKaPK0hjf.9yHxzQGZziziccp6Yng'  # The digest of 'insecure_secret'.
         sector_identifier: ''
         public: false
         redirect_uris:
           - 'https://oidc.example.com:8080/oauth2/callback'
-        audience: []
+        audience:
+          - 'unique-client-identifier'
         scopes:
           - 'openid'
           - 'groups'
@@ -94,8 +95,10 @@ configured in the application.
 This secret must be generated by the administrator and can be done by following the
 [How Do I Generate Client Secrets](../../../integration/openid-connect/frequently-asked-questions.md#how-do-i-generate-client-secrets) FAQ.
 
-This must be provided when the client is a confidential client type, and must be blank when using the public client
-type. To set the client type to public see the [public](#public) configuration option.
+This must be provided when the client is a confidential client type provided with the exception if you've configured a
+[token_endpoint_auth_method](#tokenendpointauthmethod) that uses a credential type that isn't a secret (i.e. a key), and
+must be blank when using the public client type. To set the client type to public see the [public](#public)
+configuration option.
 
 ### sector_identifier
 
@@ -157,7 +160,8 @@ their redirect URIs are as follows:
 
 {{< confkey type="list(string)" required="no" >}}
 
-A list of audiences this client is allowed to request.
+A whitelist of audiences this client is allowed to request. These audiences are automatically granted to all access
+requests by default. The client [id](#id) is automatically allowed for every client.
 
 ### scopes
