+++ internal/oidc/session.go
@@ -1,6 +1,8 @@
 package oidc
 
 import (
+	"context"
+	"net/url"
 	"time"
 
 	"github.com/google/uuid"
@@ -8,7 +10,99 @@ import (
 	"github.com/ory/fosite"
 	"github.com/ory/fosite/handler/openid"
 	"github.com/ory/fosite/token/jwt"
+
+	"github.com/authelia/authelia/v4/internal/model"
+	"github.com/authelia/authelia/v4/internal/utils"
 )
+
+// NewSession creates a new empty OpenIDSession struct.
+func NewSession() (session *Session) {
+	return &Session{
+		DefaultSession: &openid.DefaultSession{
+			Claims: &jwt.IDTokenClaims{
+				Extra: map[string]any{},
+			},
+			Headers: &jwt.Headers{
+				Extra: map[string]any{},
+			},
+		},
+		Extra: map[string]any{},
+	}
+}
+
+// NewSessionWithAuthorizeRequest uses details from an AuthorizeRequester to generate an OpenIDSession.
+func NewSessionWithAuthorizeRequest(ctx Context, issuer *url.URL, kid, username string, amr []string, extra map[string]any,
+	authTime time.Time, consent *model.OAuth2ConsentSession, requester fosite.AuthorizeRequester) (session *Session) {
+	if extra == nil {
+		extra = map[string]any{}
+	}
+
+	session = &Session{
+		DefaultSession: &openid.DefaultSession{
+			Claims: &jwt.IDTokenClaims{
+				Subject:     consent.Subject.UUID.String(),
+				Issuer:      issuer.String(),
+				AuthTime:    authTime,
+				RequestedAt: consent.RequestedAt,
+				IssuedAt:    ctx.GetClock().Now().UTC(),
+				Nonce:       requester.GetRequestForm().Get(ClaimNonce),
+				Audience:    requester.GetGrantedAudience(),
+				Extra:       extra,
+
+				AuthenticationMethodsReferences: amr,
+			},
+			Headers: &jwt.Headers{
+				Extra: map[string]any{
+					JWTHeaderKeyIdentifier: kid,
+				},
+			},
+			Subject:  consent.Subject.UUID.String(),
+			Username: username,
+		},
+		ChallengeID:           model.NullUUID(consent.ChallengeID),
+		KID:                   kid,
+		ClientID:              requester.GetClient().GetID(),
+		ExcludeNotBeforeClaim: false,
+		AllowedTopLevelClaims: nil,
+		Extra:                 map[string]any{},
+	}
+
+	// Ensure required audience value of the client_id exists.
+	if !utils.IsStringInSlice(requester.GetClient().GetID(), session.Claims.Audience) {
+		session.Claims.Audience = append(session.Claims.Audience, requester.GetClient().GetID())
+	}
+
+	session.Claims.Add(ClaimAuthorizedParty, session.ClientID)
+	session.Claims.Add(ClaimClientIdentifier, session.ClientID)
+
+	return session
+}
+
+// PopulateClientCredentialsFlowSessionWithAccessRequest is used to configure a session when performing a client credentials grant.
+func PopulateClientCredentialsFlowSessionWithAccessRequest(ctx Context, request fosite.AccessRequester, session *Session, funcGetKID func(ctx context.Context, kid, alg string) string) (err error) {
+	var (
+		issuer *url.URL
+		client Client
+		ok     bool
+	)
+
+	if issuer, err = ctx.IssuerURL(); err != nil {
+		return fosite.ErrServerError.WithWrap(err).WithDebugf("Failed to determine the issuer with error: %s.", err.Error())
+	}
+
+	if client, ok = request.GetClient().(Client); !ok {
+		return fosite.ErrServerError.WithDebugf("Failed to get the client for the request.")
+	}
+
+	session.Subject = ""
+	session.Claims.Subject = client.GetID()
+	session.ClientID = client.GetID()
+	session.DefaultSession.Claims.Issuer = issuer.String()
+	session.DefaultSession.Claims.IssuedAt = ctx.GetClock().Now().UTC()
+	session.DefaultSession.Claims.RequestedAt = ctx.GetClock().Now().UTC()
+
+	return nil
+}
 
 // Session holds OpenID Connect 1.0 Session information.
 type Session struct {
